Relocate DatastoreIO
