+++ scio-test/src/test/scala/com/spotify/scio/testing/JobTestTest.scala
@@ -17,10 +17,6 @@
 
 package com.spotify.scio.testing
 
-import java.util.Collections
-
-import com.google.datastore.v1.Entity
-import com.google.datastore.v1.client.DatastoreHelper.{makeKey, makeValue}
 import com.spotify.scio._
 import com.spotify.scio.avro.AvroUtils.{newGenericRecord, newSpecificRecord}
 import com.spotify.scio.avro._
@@ -81,16 +77,6 @@ object GenericParseFnAvroFileJob {
       PartialFieldsAvro(gr.get("int_field").asInstanceOf[Int])
     ).map(a => AvroUtils.newGenericRecord(a.intField))
       .saveAsAvroFile(args("output"), schema = AvroUtils.schema)
-    sc.run()
-    ()
-  }
-}
-
-object DatastoreJob {
-  def main(cmdlineArgs: Array[String]): Unit = {
-    val (sc, args) = ContextAndArgs(cmdlineArgs)
-    sc.datastore(args("input"), null, null)
-      .saveAsDatastore(args("output"))
     sc.run()
     ()
   }
@@ -309,33 +295,6 @@ class JobTestTest extends PipelineSpec {
     }
     an[AssertionError] should be thrownBy {
       testGenericParseAvroFileJob((1 to 4).map(newGenericRecord))
-    }
-  }
-
-  def newEntity(i: Int): Entity =
-    Entity
-      .newBuilder()
-      .setKey(makeKey())
-      .putAllProperties(Collections.singletonMap("int_field", makeValue(i).build()))
-      .build()
-
-  def testDatastore(xs: Seq[Entity]): Unit =
-    JobTest[DatastoreJob.type]
-      .args("--input=store.in", "--output=store.out")
-      .input(DatastoreIO("store.in"), (1 to 3).map(newEntity))
-      .output(DatastoreIO("store.out"))(coll => coll should containInAnyOrder(xs))
-      .run()
-
-  it should "pass correct DatastoreJob" in {
-    testDatastore((1 to 3).map(newEntity))
-  }
-
-  it should "fail incorrect DatastoreJob" in {
-    an[AssertionError] should be thrownBy {
-      testDatastore((1 to 2).map(newEntity))
-    }
-    an[AssertionError] should be thrownBy {
-      testDatastore((1 to 4).map(newEntity))
     }
   }
